BE_Refactor: Dto 구조 수정
- List Dto 타입 객체: static class로
