--- back/src/main/java/com/project/mainproject/review/dto/GetReportedReviewDto.java
@@ -1,16 +1,0 @@
-package com.project.mainproject.review.dto;
-
-import lombok.Builder;
-import lombok.Getter;
-
-import java.time.LocalDateTime;
-
-@Getter
-@Builder
-public class GetReportedReviewDto {
-    private Long reviewIdx;
-    private String reviewContent;
-    private int reportedCount;
-    private LocalDateTime reviewCreatedAt;
-
-}
