+++ back/src/main/java/com/project/mainproject/review/mapper/ReviewMapper.java
@@ -1,11 +1,12 @@
 package com.project.mainproject.review.mapper;
 
+import com.project.mainproject.review.dto.ListGetStoreReviewDto.ReplyDto;
+import com.project.mainproject.review.dto.ListGetStoreReviewDto.StoreReviewDto;
 import com.project.mainproject.review.dto.ListGetUserReviewDto.GetUserReviewDto;
 import com.project.mainproject.review.dto.ListReportedReviewDto.ReportedReviewDto;
 import com.project.mainproject.review.dto.PostCreateReviewDto;
 import com.project.mainproject.review.dto.PostUpdateReviewDto;
 import com.project.mainproject.review.dto.SimpleReviewDto;
-import com.project.mainproject.review.dto.StoreReviewPageDto;
 import com.project.mainproject.review.entity.Review;
 import com.project.mainproject.review.entity.ReviewReply;
 import org.mapstruct.Mapper;
@@ -14,8 +15,6 @@ import org.mapstruct.MappingTarget;
 import org.mapstruct.ReportingPolicy;
 
 import java.util.List;
-
-import static com.project.mainproject.review.dto.StoreReviewPageDto.ReplyDto;
 
 @Mapper(componentModel = "spring", unmappedTargetPolicy = ReportingPolicy.IGNORE)
 public interface ReviewMapper {
@@ -29,7 +28,7 @@ public interface ReviewMapper {
     @Mapping(target = "storeIdx", source = "store.storeIdx")
     SimpleReviewDto reviewToSimpleReviewDto(Review review);
 
-    List<StoreReviewPageDto> reviewsToReviewsDto(List<Review> reviews);
+    List<StoreReviewDto> reviewsToReviewsDto(List<Review> reviews);
 
     @Mapping(target = "reviewImage", expression = "java(review.getReviewImages().size() == 0 " +
                                                         "? null " +
@@ -38,7 +37,7 @@ public interface ReviewMapper {
     @Mapping(target = "userIdx", source = "user.userIdx")
     @Mapping(target = "userName", source = "user.name")
     @Mapping(target = "profileImage", source = "user.imagePath")
-    StoreReviewPageDto reviewToReviewDto(Review review);
+    StoreReviewDto reviewToReviewDto(Review review);
 
     @Mapping(target = "userIdx", source = "user.userIdx")
     @Mapping(target = "userName", source = "user.name")
